perf: replace renderer draw mutex with spinlock
This lock is rarely contended, and when it is the contention will likely
be very very short, so even the overhead of a syscall (much less the
overhead of falling asleep if there is contention) is way bigger than
had ought to be.

It's hard to measure an improvement for this in benchmarks, but I'm
fairly confident this is more efficient and has a tiny bit less latency
over all compared to before, if only for the removal of the syscall.